requirements.txt
fastapi==0.116.0
matplotlib==3.10.3
mne==1.9.0
numpy==2.3.1
pandas==2.3.1
seaborn==0.13.2
SQLAlchemy==2.0.41
uvicorn==0.35.0
gunicorn
dotenv
python-jose
passlib